Remove the remaining word-size dependencies
None of these were wrong on LP64, but each meant something different
where long is 32 bits.

CreateBVConst chopped its uint64 argument into chunks the width of a
machine word. That happened to be correct on both word sizes, but only
because Chunk_Store also takes an unsigned long, and it needed a
-Wshift-count-overflow suppression for a shift that is dead on 64-bit
and live on 32. Store 32 bits at a time instead, which is what CBVOps
already does for the same reason, and the coupling and the pragma both
go away.

Checked against an independent oracle over widths 1 to 130 -- either
side of both chunk boundaries -- for boundary values and for every
single-bit value: 10400 cases, all agreeing. Deliberately mis-chunking
the loop fails 3294 of them, so the check is not vacuous.

MersenneTwister typedef'd its state word to unsigned long, so the state
array was twice the necessary size on LP64 and the /dev/urandom seed
read 8 bytes per word to then discard 4. The algorithm masks every state
update to 32 bits -- initialize() and twist() both -- so the generated
sequence is unchanged; the seeded exhaustive tests confirm it.

test_fprewrites computed a symbol's domain as "1ul << width", which is
fine on LP64 and undefined where long is 32 bits, and only checked the
enumeration limit after the product was formed. Saturate the shift and
test the limit before each multiply, so neither can overflow.

The rest is spelling: loop counters and widths that were long or
unsigned long for no reason.

+++ lib/Simplifier/RemoveUnconstrained.cpp
@@ -1060,7 +1060,7 @@ ASTNode RemoveUnconstrained::topLevel_other(const ASTNode& n,
           newWidth += rhsSize;
         }
 
-        assert(newWidth == (long int)operandWidth);
+        assert(newWidth == (int)operandWidth);
         replace(var, current);
       }
       break;
